requirements.txt
# FastAPI and server
fastapi==0.115.0  # Updated for Python 3.13 compatibility
uvicorn[standard]==0.32.0  # Updated for Python 3.13 compatibility
pydantic==2.10.0  # Updated for Python 3.13 compatibility
pydantic-settings==2.6.0  # Updated for Python 3.13 compatibility
python-multipart==0.0.12  # Updated
email-validator==2.1.1  # Required for EmailStr validation

# Database
sqlalchemy==2.0.36  # Updated
alembic==1.14.0  # Updated
asyncpg>=0.30.0  # Updated for Python 3.13 compatibility (use latest)
psycopg2-binary>=2.9.9  # Required for Alembic migrations (sync connection)
greenlet>=3.0.0  # Required for SQLAlchemy async operations

# Authentication and Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
bcrypt==4.1.2

# Blockchain (Optional - has Python 3.13 compatibility issues)
# bigchaindb-driver==0.6.2  # BigchainDB Python driver (deprecated, incompatible with Python 3.13)
# Note: BigchainDB dependencies (PyNaCl, pysha3, python-rapidjson) don't support Python 3.13
# The system will use mock transactions if BigchainDB driver is not available

# HTTP client
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
aiosqlite==0.19.0  # For in-memory SQLite testing

# Logging
python-json-logger==2.0.7

